requirement.txt
torch==1.6.0
torchvision==0.7.0
dgl-cu102==0.5.2
tqdm
pandas
rdflib